Add native Windows support to build.sh
The Windows port to build may be specified using the WINPORT environment
variable to select mingw, mingw64, msvc or msvc64. If no gcc is found,
the script will determine a WINPORT to use if it can find a C compiler
for one (MSVC is preferred followed by mingw64 then mingw32).